requirements.txt
torch
numpy
albumentations
pillow
scikit-learn
PyYAML
ultralytics
opencv-python
langchain